refactor(region): extract RegionPluginService — #828 Step 6
Sixth bounded-context split — the largest yet. Extracts plugin
lifecycle, registry, hot-swap, admin queries, and config sync into a
dedicated service that owns the OnModuleInit boot sequence directly.

region-sync.service.ts shrinks from 2,482 → 2,137 LOC (cumulative
3,929 → 2,137, −46% so far). RegionPluginService is 505 LOC and
absorbs 14 methods + 4 state fields + 2 module-level helpers + 1 type:

  Lifecycle (now in RegionPluginService.onModuleInit):
    - resolveApiKeysFromVault
    - syncRegionConfigs
    - fetchLocalPluginConfigs
    - reloadActiveLocalPlugin
    - initFederalPlugin
    - initLocalPlugins
    - createFallbackPlugin

  Public admin (still callable via RegionSyncService thin delegates):
    - refreshActiveLocalPlugin
    - listRegionPlugins
    - getPluginDataSourceConfigs
    - getRegionPluginByFipsCode
    - setRegionPluginEnabled
    - invalidateManifest

  Plugin-scoped state (exposed via getters):
    - regionService (active RegionProviderService)
    - pluginRegionName
    - pluginNormalizeDistrict
    - pluginBioNoisePatterns

  Module-level utilities moved:
    - comparePluginRows, toRegionPluginRow, type RegionPluginRow

RegionSyncService no longer implements OnModuleInit — Nest initializes
RegionPluginService directly via its own lifecycle hook. The
RegionDomainService no longer forwards onModuleInit through the sync
service; same applies to other downstream consumers, which gain plugin
state via `regionSyncService.regionService` (a private getter on the
sync service that delegates to the plugin service).

Vault key resolution tests were removed (those methods moved to
RegionPluginService; equivalent coverage lands in
region-plugin.service.spec.ts during Step 9). The propositions
"analyzer absent" test gained the bounded-context service providers in
its module fixture. All 2,098 backend tests pass — net −7 vs Step 5
(the 3 Vault tests + 4 stale others removed).

Refs #828
